LAB-2 files
The above screenshots are from an experiment with the RISC-V toolchain. We use riscv-gcc to compile a C-program to compute the sum of numbers from  1 to n. To minimize the code length and improve the performance, we code the addition logic in the assembly. That is stored in the load.s file. We compile the program with the load.s and verify the object dump using the objdump tool.

The output is run using the spike debugger tool and this can be used to debug the registers at intermediate stages.