Fetch and Decode
This file consists of the TL-Verilog code for the fetch and decodes logic of the CPU. The fetch logic makes use of the PC logic from the earlier commit. The instructions are stored in a memory macro provided, and the instructions are fetched from memory by connecting the memory ports to the required pins. 

Once the instruction is fetched from memory, it is decoded into various fields, based on the RISC-V ISA specs. The opcode, funct3, funct7 fields help realize the functions to be done by the ALU logic. rs1,rs2, rd fields specify the source and destination registers, which will function as the index in the register file write or read. 

The immediate field is used in almost many operations, and the immediate value is encoded in the instructions, in different ways, based on the type of instruction.