Register File read
We make use of a register file comprising of 32 registers of 32 bits (as XLEN = 32). This register file supports two reads and one write, pertaining to the RS1, RS2 requirements from the RV32I set. 

Based on the rs1,rs2 indexes from the instructions, we read the register values and provide it to the ALU for computation. The reset values for each register will be its index.