Register Bypass - Solving Read after write Hazard
Implementing pipelines always comes with hazards. There are two common hazards, Read After Write Hazard and Control Flow hazard.

Read After Write Hazard occurs whenever there is an instruction requesting the read of a register just after an instruction that executed a write into the same register. 

So, this is taken care using the register bypass, where we add muxes at the output of the register file, on the read data outputs. If the source register index of the two instructions match, the ALU output is selected. Else, the read data is sent through.