Branch instruction - PC update
Whenever the branch instruction is encountered, the branch computation logic calculates the jump address and allocates it to the next PC logic. But the issue with a pipelined CPU is that, the branch computation will happen every cycle because, it is just an addition logic with the PC and the immediate value. So, to avoid this, we need to invalidate the instructions succeeding the branch and update the PC correctly. Since, the next PC/Prefetch logic sits at the first stage, we need to provide the branch computation logic, which is computed at the Execute stage and assign it to the PC correctly. So, in other words, the PC should take the target address at the third instruction after the branch is fetched. So, the valid signal which is pulsed once in three cycles, will help us to invalidate the instructions issued after the branch.