Complete Instruction set decode and ALU
This version of the code comprises of the complete instruction set decoding and the corresponding ALU operation for each instruction. Also, to ensure that the CPU operates on an instruction every cycle, we have modified the valid logic to ensure that the instructions are valid except for the two instructions issued after the branch. 

Concurrently, the next PC logic is updated to reflect this change.