Load and Store
This version consists of the load and store instructions implemented with the CPU. For loading and storing the data from the ALU, we decode the corresponding instructions and use them. For load, only the lw instruction is supported. This will load the entire 32 bits into the data memory. The data memory used is a macro present in Makerchip, coded using TL-Verilog. 

The load instruction will face a similar hazard like the case of read after write with the register file. So, we need to invalidate the next two instructions after the load instruction to ensure the execution is correct and valid. So, the valid condition checks for the issue of load instruction in the previous two cycles and decides whether the current instruction is valid. This implementation is similar to our branch instruction implementation.

Store is straightforward in comparison to the load instruction. The rd index is used as the address and the enable is asserted if there is a valid store instruction. 

In terms of instruction pre-fetch, the PC needs to keep track of the address of the instruction next to the load. This is achieved using the pipeline flops for the next pc calculation. If there is a valid load instruction in the execute stage, the PC has to be loaded with the incremented address present at the fetch stage of the load.